# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PDF Transaction WebAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/transaction-category": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Get transaction categories",
                "description": "Get transaction categories",
                "operationId": "get_transactionCategories",
                "responses": {
                    "200": {
                        "description": "Successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Not Authorized"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Create transaction category",
                "description": "Create transaction category",
                "operationId": "create_transactionCategory",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "transaction_type_id",
                                    "name",
                                    "offset",
                                    "customer",
                                    "sender"
                                ],
                                "properties": {
                                    "transaction_type_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "integer"
                                    },
                                    "offset": {
                                        "type": "text"
                                    },
                                    "customer": {
                                        "type": "boolean"
                                    },
                                    "sender": {
                                        "type": "boolean"
                                    },
                                    "descriptions[]": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Not Authorized"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    }
                }
            }
        },
        "/api/transaction-category/{id}": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Get transaction category",
                "description": "Get transaction category",
                "operationId": "get_transactionCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "transaction category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Not Authorized"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Update transaction category",
                "description": "Update transaction category",
                "operationId": "update_transactionCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "transaction category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "transaction_category_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "text"
                                    },
                                    "offset": {
                                        "type": "integer"
                                    },
                                    "customer": {
                                        "type": "boolean"
                                    },
                                    "sender": {
                                        "type": "boolean"
                                    },
                                    "descriptions[]": {
                                        "type": "text"
                                    },
                                    "descriptions_to_delete[]": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Not Authorized"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Delete transaction category",
                "description": "Delete transaction category",
                "operationId": "delete_transactionCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "transaction category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Not Authorized"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    }
}